+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/model/SubtaskTemplateType.java
@@ -1,0 +1,9 @@
+package de.codescape.jira.plugins.multiplesubtasks.model;
+
+public enum SubtaskTemplateType {
+
+    USER,
+    PROJECT,
+    GLOBAL
+
+}